// src/app/models/Cronograma.ts
export class Cronograma {
  precioVehiculo: number = 0;
  marcaVehiculo: string = '';
  modeloVehiculo: string = '';
  tipoMoneda: string='';
  numeroAnios: number = 0;
  porcentajeCuotaInicial: number = 0;
  tipoTasaInteres: string = '';
  plazoTasaInteres: string = '';
  porcentajeTasaInteres: number = 0;
  capitalizacion: string = '';
  plazoDeGracia: string = '';
  tiempoPlazoDeGracia: number = 0;
  porcentajeSeguroDesgravamen: number = 0;
  tiempoSeguroDesgravamen: string = '';
  porcentajeSeguroVehicular: number = 0;
  tiempoSeguroVehicular: string = '';
  portes: number = 0;
  costosRegistrales: number = 0;
  costosNotariales: number = 0;
  frecuenciaPago: string = '';
  fechaInicio: string = '';
  porcentajeCuotaFinal: number = 0;

  constructor() {
  }
}